Password without rar,zip,gz extract and zip adding
* 1. öncelikli rar, zip, gz extractor olarak kullanılır. (Dosyaların bunun bulunduğu klasöre taşınması gerekir daha sonra arayüzden seçilerek extract edilmesi sağlanır.)
* 1. öncelikli zip dosyaları oluşturma için kullanılır.
---------------------------------------------------------------------------------------------------------------------------------------------
# The Unzipper

The Unzipper extracts .zip and .rar archives or .gz files on webservers. It detects .zip/.rar/.gz archives and let you choose which one to extract (if there are multiple archives available).
As of version 0.1.0 it also supports creating archives.

It's handy if you do not have shell access. E.g. if you want to upload a lot of files (php framework or image collection) as archive - because it is much faster than uploading each file by itself.